crawl/flask/requirements.txt
aniso8601==10.0.0
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
ConfigArgParse==1.7
exceptiongroup==1.2.2
Flask==3.0.0
Flask-Cors==5.0.0
Flask-Injector==0.15.0
Flask-Login==0.6.3
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
gevent==24.11.1
geventhttpclient==2.3.3
google-play-scraper==1.2.4
googlesearch-python==1.3.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_resources==6.5.2
injector==0.22.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
locust==2.32.5
lxml==5.3.1
MarkupSafe==3.0.2
msgpack==1.1.0
numpy==2.2.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
psutil==6.1.1
pycparser==2.22
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
tomli==2.2.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
